💄 Added stilying to send report, send review and added slider for rating

+++ backend/HotelManagement.Core/Reviews/CreateReviewCommand.cs
@@ -8,7 +8,7 @@ public record CreateReviewCommand(
     Guid PropertyId, 
     Guid UserId,
     string Title,
-    int Rating,
+    double Rating,
     string Description
 ) : ICommand<Guid?>;
